Scripts/Battlefield: Move npc_wg_give_promotion_credit script to BattlefieldWG
The script references BattlefieldWG, which is not exported in dynamic builds.
Move the script, so it is in the same library.

(cherry picked from commit c40b897d369cf45696ca626fdf48bf51226a73a8)
